Use a standard icon for "Link" button
The icon named "link" is not a standard name from the freedesktop.org
Icon Naming Specification. Hence it may not be available on the user's
system, giving an unappealing blank button in the interface.

The icon "go-jump" is the standard icon for a "jump to action", which is
appropriate to the function of the button.